Add an `IntoGenerator` trait
This mirrors `IntoIterator`.  This is needed for `flatten` / `flat_map`
support.

In addition to `impl IntoGenerator for Generator`, this provides Vec/Slice
impls to make example code work better.  We can mirror all of
the `IntoIterator` impls over time.

Ideally, we'd have `impl IntoGenerator for IntoIterator` but we can't
have two blanket impls.  This would have avoided the need for all of the
impls.  The plus side is that this makes native/fast impls easy and
slow/adapted impls possible with `from_iter`.