Add `from_iter` for interoperability
My original plan had been for `IntoGenerator` to be implemented for both
`Generator` and `IntoIterator` but that fell apart because you can't
have two blanket impls for a trait, so I prioritized `Generator` for
easier interop for people fully into the `pushgen` ecosystem.

This still left my `IntoIterator` work, so I broke it out into
`from_iter`, which is like `from_fn`.  This should still make it
relative easily to interop with iterators.  We'll just need to add more
concrete implementations for `IntoGenerator`.

The other benefit of this split is we won't be forced into delegated
`IntoGenerator` to `IntoIterator` for slices, allowing us to use the
native `SliceGenerator`.